requirements.txt
fastapi
uvicorn
httpx
jinja2
python-dotenv
pydantic
python-multipart
aiofiles
firebase-admin
pydantic-settings
openpyxl
pandas
fuzzywuzzy
python-Levenshtein
